virtcontainers: Force the hypervisor plugin to be used
Because the goal here is only to show a point that an hypervisor
plugin could work, there is no need to implement the whole proper
path to configure a PluginHypervisor type. Instead, this commit
is a hack forcing this hypervisor implementation to be picked up,
and using a specific path for the plugin.

Fixes #51

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>