virtcontainers: Introduce containers map
The purpose of such a mapping is to quickly provide to the caller of
the API the list of sandboxes including a specific container ID.

This will be used as an optimization by the caller in order to avoid
calling ListPod() which costs a lot since it has to open every single
pod and container files on the filesystem to return the mapping.

This commit allows for a complexity reduction from O(n²) to O(1).

Fixes #212

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>