virtcontainers: Update containers map only if stateful
Every time a container is created or deleted, virtcontainers has to
update the container/sandbox mapping. That being said, some consumers
of virtcontainers might keep their own tracking since they are some
long living process able to do so. In this case, the use of a new flag
"Stateless" through the sandbox configuration let virtcontainers know
that it does not need to update the container mapping. This is for
performance purpose, in order to prevent the code from unnecessary
filesystem accesses.

Fixes #212

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>