cli: Optimize container research relying on ContainerSandboxList()
This commit leverages the new function ContainerSandboxList() added
to the virtcontainers API. This will allow for better performance
regarding the time spent to retrieve the sandbox ID related to a
container ID.

This commit also updates the whole unit tests as a consequence. This
is involving most of them since they were all relying on ListPod()
before.

Fixes #212

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>