virtcontainers: Remove RunSandbox() from API
First of all, RunSandbox() was never used since an OCI "run" was
actually relying on CreateSandbox() + StartSandbox(), meaning there
was no consumer of this function.

Second, and that's the main reason here, with the refactoring of the
network, by moving part of the network creation up to the CLI, a big
function like RunSandbox() cannot exist. The granularity is not thin
enough, which would end up setting up the network after this function
returns, which would be wrong as the sandbox and its containers would
be running already.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>