cli: Agent as init in rootfs images
Let kernel start the default init daemon (/sbin/init) and rid of all related to
systemd.
Currently we are tied to systemd because we have hardcoded it, we use
systemd even when we don't know if the image has it. Now that we can use
kata-agent as init process in rootfs images (kata-containers/agent#376), we
should use it, since we are responsible for any security flaw found in
kata-containers, it doesn't matter if we wrote that code or not (systemd).
The usage of kata-agent as init process brings performance improvements in
boot time and memory footprint.

Depends-on: github.com/kata-containers/tests#948

fixes #767

Signed-off-by: Julio Montes <julio.montes@intel.com>