virtcontainers: should be able to update created container
Original logic would not allow a container to be updated if it wasn't
in the running state. A container should be able to be updated if it
is either running or is created and in ready state.

This better matches the logic in CRI-O today, and allows for the use
case where a created container is updated by due to K8S CPU manager
adjusting the CPU affinity of a created but not yet running container.

Fixes #878

Signed-off-by: Eric Ernst <eric.ernst@intel.com>