fix: fail to do clean without named return
```
defer func() {
    if err != nil {

    }
}
```
when returned variable is not named, if err is re-defined in if
condition, like `if err := s.getAndStoreGuestDetails(); err != nil`
in `createSandboxFromConfig()`, defer function can not catch error.
So it better to make returned err named.
Other change in virtcontainers/sandbox.go and virtcontainers/vm.go
is not must, but make more sense.

Fixes #1076

Signed-off-by: Ace-Tang <aceapril@126.com>