virtcontainers: Jailer: Add jailer support for firecracker
Firecracker provides a jailer to constrain the VMM. Use this
jailer to launch the firecracker VMM instead of launching it
directly from the kata-runtime.

The jailer will ensure that the firecracker VMM will run
in its own network and mount namespace. All assets required
by the VMM have to be present within these namespaces.

This also means that more sandbox context needs to be made
available to the hypervisor/jailer.

If a jailer is not available for a given hypevisor the
runtime will act as the jailer.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>