shimv2: Close vhostfd after vm get vhostfd
If kata containers is using vfio and vhost net，the unbinding
of vfio would be hang. In the scenario, vhost net kernel thread
takes a reference to the qemu's mm, and the reference also includes
the mmap regions on the vfio device file. so vhost kernel thread
would be not released when qemu is killed as the vhost file
descriptor still is opened by shim v2 process, and the vfio device
is not released because there's still a reference to the mmap.

Fixes: #1669

Signed-off-by: Yang, Wei <w90p710@gmail.com>
Signed-off-by: Eric Ernst <eric.ernst@intel.com>
(cherry picked from commit 071030b7846e765ffa3db1ee10c29552f383cbb3)
Signed-off-by: Ganesh Maharaj Mahalingam <ganesh.mahalingam@intel.com>