network: Detect and add static ARP entries
Some network plugins add static arp entries in the network
namespace.
Scan namespace for static entries and pass these on to the
agent to be added within the guest.
If the grpc api is not implemented by the agent due to a older running
agent, check for this and do not error out to maintain
backward compatibility.

Fixes #2671

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>