clh: Remove the use of deprecated '--memory file=' parameter
Along with the release of cloud-hypervisor v0.8.0, this option has been
deprecated. clh now enforces to use the alternative controls,
e.g. "shared" and "hugepages", which can infer the backing file
paths. Also, we don't use "hugepages" in kata, so we are fine now as the
"shared" control  is already enabled.

Fixes: #2803

Signed-off-by: Bo Chen <chen.bo@intel.com>