kata_agent: Revert: Allow to use "VirtPath" as volume source for blk device
This reverts commit aea29b64b66f75049cb045f9e41dff2becdbebdc.
The above commit was a temporary (not reliable) workaround to support
container rootfs through hotplugged block device when using
cloud-hypervisor (clh). As the updated version of clh now returns the
PCI BDF information for all hotplugged devices, we no longer need to
rely on the predicted device path in the guest (a.k.a 'VmPath'). This
patch reverts the above commit to remove the special code path for
supporting cloud-hypervisor.

Fixes: #2908

Signed-off-by: Bo Chen <chen.bo@intel.com>