clh: Set the BDF info of hotplugged blk devices as the drive PCIAddr
This patch sets the 'drive.PCIAddr' with the BDF info when hotplugging
block devices, and also sets the 'drive.Virtpath' as 'nil'
explicitly. In this way, the kata-agent will no longer rely on the
predicted  device path in the guest to locate the block device and
volumes, instead it will rely on the PCI BDF of the hotplugged block
device.

Depends-on: github.com/kata-containers/agent#824

Fixes: #2908

Signed-off-by: Bo Chen <chen.bo@intel.com>