clh: openapi: Tag the 'openapi-generator-cli' container to v4.3.1
This patch ensures that we can generate the client code based on
cloud-hypervisor's openAPI spec consistently across different systems.
The v4.3.1 tag is the latest stable release of the openapi-generator-cli
container. This patch also re-generated the client code with the tagged
container image, which should have no functional changes on the generated
client code.

Signed-off-by: Julio Montes <julio.montes@intel.com>
Signed-off-by: Bo Chen <chen.bo@intel.com>