block: Use PciPath type through block code
BlockDrive::PCIAddr doesn't actually store a PCI address (DDDD:BB:DD.F) but
a PCI path.  Use the PciPath type and rename things to make that clearer.

TestHandleBlockVolume() previously used a bizarre value "0002:01" for the
"PCI address" which was neither an actual PCI address, nor a PCI path.
Update it to use a PCI path - the actual value appears not to matter in
this test, as long as its consistent throughout.

fixes #3002

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>