vhost-user-blk: Use PciPath type for vhost user devices
VhostUserDeviceAttrs::PCIAddr didn't actually store a PCI address
(DDDD:BB:DD.F), but rather a PCI path.  Use the PciPath type and rename
things to make that clearer.

TestHandleBlockVolume previously used the bizarre value "0001:01" which is
neither a PCI address nor a PCI path for this value.  Change it to a valid
PCI path - it appears the actual value didn't matter for that test, as long
as it was consistent.

fixes #3002

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>