sandbox: add support for shared directory for observability logs
We'd like to be able to reliably pull 'custom userspace application'
logs from the Kata sandbox to facilitate getting audit-like event details
out of the guest. By creating a /var/log/<sandbox id> directory on the
host and bind mounting it to the host shared directory, we can provide a
path that custom guest-userspace applications can leverage to provide
logs onto the host.

This path and mount is tied to the lifecycle of the sandbox, and is not
exposed to the underlying container. Filesystem sharing is required for
this.

Fixes: #3137

Signed-off-by: Eric Ernst <eric.g.ernst@gmail.com>